Adding the first version of the testsuite tool
The testsuite tool will be used for testing future tools. This
commit also introduces a directory structure for the project:

  - /bin   Executable files.
  - /lib   Library function files.
  - /test  Tests for the libraries.

Apart from the /test folder, the project tries to follow the FHS
standard.

Library files contain only functions which should be prefixed
by the file name. Executable files point to those libraries and
expose them as a command.