Coding style and cleanup. Minor output changes.
Most of the code is now under 78 columns of text, formatted
to be read using 8-space hard tabs.

Hard tabs are required for shell indented heredocs, so in our
case isn't much of a choice. The 8-space is the most popular
configuration (also the one GitHub has chosen), but this
might change in the future.

The trix tool now redirects everything the setup and clean
phases of a text matrix does to the sterr.