Improvements on posit, fixing timeout command issues, error handling
The posit/cov.sh was changed to pass the pattern for test files
to depur --ignore. This will prevent tests for --report=cov to
display results for the test files themselves.

posit now changes the errexit option (set +e) before running a test,
but it reverts back to whatever the setting was just after running
it. This will prevent some shells from interrupting the execution
of the entire suite.

The timeout command has differences on various platforms. A real
time test to check for the most common versions was added and
it should not break on busybox anymore.

Several instances of unquoted variables and other misc coding
formatting bits were applied.